requirements.txt

annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.8
filelock==3.18.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
platformdirs==4.3.7
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
starlette==0.45.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
